requirements.txt
Flask==1.0.2
google-cloud-vision==0.36.0
Pillow==6.0.0
